Partial namespacing and refactor of common Ruby scripts.
