--- libproject/workbench-common.rb
@@ -1,1 +1,0 @@
-WORKBENCH_ROOT = File.expand_path(File.join(File.dirname(__FILE__), '..'))
